move the family location activities to address module
