+++ Modules/Marriage/Register/Marriage/RegisterValid/VerifyHusband.php
@@ -23,7 +23,7 @@ class VerifyHusband
         if($user->profile()->data_of_birth >= 567648000){
             return true;
         }else{
-            $this->error = ["$user->first_name $user->last_name was too early to marry"];
+            $this->error = ["family authentication fails too early to marry"];
         }
     }
 
@@ -36,14 +36,14 @@ class VerifyHusband
         if($wife < 4){
             return true;
         }else{
-            $this->error = ["$user->first_name $user->last_name you can not marry more than 4 wives"];
+            $this->error = ["family authentication fails can not marry more than 4 wives"];
         }
     }
     
     public function validBirth(User $user, $data)
     {
         if($user->profile()->child()->birth()->mother_id != Wife::where('status_id',$data['mstatus'])->get()->mother()->id){
-            $this->error = ["Invalid husband and his mother information"];
+            $this->error = ["husband birth authentication fails and his mother information"];
         }
     }
